requirements.txt
boto3==1.4.6
botocore==1.6.0
certifi==2017.7.27.1
chardet==3.0.4
docutils==0.14
futures==3.1.1
idna==2.5
jmespath==0.9.3
pluggy==0.4.0
py==1.4.34
python-dateutil==2.6.1
python-gitlab==0.21.2
requests==2.18.3
s3transfer==0.1.10
six==1.10.0
tox==2.7.0
urllib3==1.22
virtualenv==15.1.0
